refactor(commons): group the legacy migration scripts under prisma/legacy-migration
- rename archive.ts, patch.ts and archive-upload.ts to initial-import.ts,
  apply-diff.ts and upload-files.ts, naming the job rather than the source
- move the staged output and avatar snapshot under the same directory
- run db:legacy:patch through tsx directly; wrapping it in `sh -c` swallowed
  the --apply / --verify-only / --skip-upload flags it reads from argv
