+++ apps/modeling-commons-backend/package.json
@@ -37,8 +37,8 @@
     "db:migrate": "prisma migrate deploy",
     "db:migrate:dev": "prisma migrate dev",
     "db:seed": "prisma db seed",
-    "db:seed:archive": "sh -c 'PRISMA_SEED_FILE=archive.ts yarn run db:seed'",
-    "db:patch": "tsx prisma/patch.ts",
+    "db:legacy:import": "sh -c 'PRISMA_SEED_FILE=legacy-migration/initial-import.ts yarn run db:seed'",
+    "db:legacy:patch": "tsx prisma/legacy-migration/apply-diff.ts",
     "db:reset": "prisma migrate reset --force",
     "deps:validate": "depcruise src --config .dependency-cruiser.cjs --output-type err-long",
     "deps:graph": "depcruise src --include-only \"^src\" --config --output-type dot | dot -T svg > doc/dependency-graph.svg"
