+++ apps/modeling-commons-backend/prisma/legacy-migration/lib/normalize.ts
